+++ src/main/java/crystalspider/harvestwithease/config/HarvestWithEaseConfig.java
@@ -50,6 +50,15 @@ public class HarvestWithEaseConfig {
 	}
 
   /**
+   * Returns the value of {@link CommonConfig#grantedExp}.
+   *
+   * @return {@link CommonConfig#grantedExp} as read from the {@link #COMMON common} configuration file.
+   */
+  public static Integer getGrantedExp() {
+		return COMMON.grantedExp.get();
+	}
+
+  /**
    * Common Configuration for Harvest with ease.
    */
   public static class CommonConfig {
@@ -66,6 +75,11 @@ public class HarvestWithEaseConfig {
      * Effective only if greater than 0 and {@link #requireHoe} is true.
      */
     private final ConfigValue<Integer> damageOnHarvest;
+    /**
+     * Amount of experience to grant on harvest.
+     * Effective only if greater than 0.
+     */
+    private final ConfigValue<Integer> grantedExp;
 
     /**
      * Defines the configuration options, their default values and their comments.
@@ -76,6 +90,7 @@ public class HarvestWithEaseConfig {
 			crops = builder.comment("List of in-game IDs of additional crops").define("crops", new ArrayList<String>());
 			requireHoe = builder.comment("Require holding a hoe (either hands) to right-click harvest").define("require hoe", false);
 			damageOnHarvest = builder.comment("If [require hoe] is set to true, damage the hoe of the given amount (0 to disable, must be an integer)").define("damage on harvest", Integer.valueOf(0));
+			grantedExp = builder.comment("Amount of experience to grant on harvest (0 to disable, must be an integer).").define("exp on harvest", Integer.valueOf(0));
 		}
 	}
 }