+++ CHANGELOG.md
@@ -8,6 +8,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 - Nothing new
 
+## [1.3.0] - 2022/04/15
+- Added a new configuration option: exp.  
+This new option allows to define a custom experience value to grant the player upon harvesting.  
+It must be an integer value and will be effective only if greater than 0.
+Note that this value is in experience points and not experience levels.
+
 ## [1.2.0] - 2022/04/14
 - Added a new configuration option: damage on harvest.  
 This new option allows to define a custom damage value to apply on the hoe used to right-click harvest.  
@@ -52,6 +58,7 @@ It must be an integer value and will be effective only if greater than 0 and req
 
 [Unreleased]: https://github.com/Nyphet/harvest-with-ease
 [README]: https://github.com/Nyphet/harvest-with-ease#readme
+[1.3.0]: https://github.com/Nyphet/harvest-with-ease/releases/tag/v1.3.0
 [1.2.0]: https://github.com/Nyphet/harvest-with-ease/releases/tag/v1.2.0
 [1.1.4]: https://github.com/Nyphet/harvest-with-ease/releases/tag/v1.1.4
 [1.1.3]: https://github.com/Nyphet/harvest-with-ease/releases/tag/v1.1.3